+++ app/src/main/java/se/nackademin/examensarbete/buildings/LumberMill.java
@@ -8,9 +8,9 @@ public class LumberMill implements Building {
     private float woodResourcePerMinute = 2f;
     private float stoneResourcePerMinute = 0f;
     private float catResourcePerMinute = 0f;
-    private int catCost = 50;
-    private int stoneCost = 0;
-    private int woodCost = 0;
+    private float catCost = 50;
+    private float stoneCost = 0;
+    private float woodCost = 0;
 
     public LumberMill() {
 
@@ -22,17 +22,17 @@ public class LumberMill implements Building {
     }
 
     @Override
-    public int getCatCost() {
+    public float getCatCost() {
         return catCost;
     }
 
     @Override
-    public int getWoodCost() {
+    public float getWoodCost() {
         return woodCost;
     }
 
     @Override
-    public int getStoneCost() {
+    public float getStoneCost() {
         return stoneCost;
     }
